REFACTOR: Move Markov categories to new submodule `Experimental`.
