+++ src/Catlab.jl
@@ -5,5 +5,6 @@ include("doctrines/Doctrines.jl")
 include("wiring_diagrams/WiringDiagrams.jl")
 include("graphics/Graphics.jl")
 include("programs/Programs.jl")
+include("experimental/Experimental.jl")
 
 end
