+++ test/experimental/Experimental.jl
@@ -1,0 +1,5 @@
+using Test
+
+@testset "MarkovCategories" begin
+  include("Markov.jl")
+end
